requirements.txt
# TradeHub 依赖

# 飞书集成（Phase 3+ 再启用）
lark_oapi>=1.0.0

# 环境变量
python-dotenv>=1.0.0

# Web API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pyyaml>=6.0

# 结构化解析（Phase 2-3）
# instructor>=1.0.0
# pydantic>=2.0.0  # fastapi 已包含
